allow use of continuum when fitting detectorMaps
This applies to both FitDistortedDetectorMapTask and AdjustDetectorMapTask.

Continuum contains a lot of useful information that we can use in fitting
a detectorMap, especially when arc lines are sparse (e.g., at the blue
end of the blue arm). Continuum measurements made by the CentroidTracesTask
are converted to arc line measurements with a guess wavelength and a large
wavelength error (tracesToLines). We need to be careful that the large
number of continuum measurements (one per fiber per row) doesn't dwarf the
importance of the bona fide arc line measurements, especially for the
measurement of the wavelength distortions.

We allow extra weighting factors (besides the measurement errors), which
can be assigned by the line description (e.g., downweight the continuum
measurements by giving "Trace" a weight of 10^-3). But we found this wasn't
as important as carefully excluding continuum measurements from wavelength-
related fits, and introducing an extra iteration in which the fake
wavelengths of the continuum measurements are updated.

Added an extra debugging plot of the continuum measurements, and extra
quality measurements (split sample by line description, and take a sample
of fibers.

Increased the default fitting order from 4 to 7 after noticing some wiggles
in the continuum position residuals.