interpolateFlux: correct for Jacobian when resampling
This corrects for differences between the input and output wavelength
spacing, which is important for getting the continuum level reasonable.
Added an option to disable the correction, and another to indicate whether
the correction should be squared (for variances).

Need to take care that the correction is only applied once, and that we
maintain the 'flux' array in units something close to counts. Since the
normalisation carries the absolute flux level that we care about, we only
apply it to that.

Moved the interpolate module into pfs.drp.stella (from ...datamodel)
because it's more general purpose than just applying to the datamodel.