mergeArms: normalise fibers independently
We've been adopting a common normalisation for all fibers, but that doesn't
ensure all fibers have flux values close to counts. Instead, calculate the
normalisations for fibers independently, sum the normalised fluxes in the
overlap, and calculate the sum of the normalisations.

This change also forces the normalisation to zero outside the measured
wavelengths (rather than having it go strongly negative near the edges).