PfsFiberArray(Set): add jacobian option to resample methods
Most interpolation in the pipeline uses the PfsFiberArray.resample
and PfsFiberArraySet.resample methods. Give those the option to
apply the Jacobian or not. The default for PfsFiberArraySet.resample
is jacobian=True, because most often we resample from extracted
CCD spectra, where the wavelength sampling is important. The default
for pfsFiberArray.resample is jacobian=False, because we don't
care about the sampling.