add SelectFibersTask
This is a relatively simple function that doesn't need to be a Task
except that that's a common way to associate a Config and simplify
reuse. The Task allows selection of fibers by fiberStatus and
targetType, with an optional odd/even filtering. It was adapted
from the DummySubtractSky2dTask, and is now also applied to the
SubtractSky2dTask so the user can easily control the selection of
fibers for sky subtraction without having to hack the pfsConfig.