reduceExposure: update lines after photometry
The ArcLineSet we pass to the PhotometerLinesTask already
has intensity measurements, but they don't match the
NevenPsf that gets installed and used for the photometry.
We therefore need to clobber the old lines measurements
and replace it with the lines from the photometry, so
when we come to subtractSky2d, the PSF and the photometry
match.