FiberProfileSet: require identity at construction
In order to write fiberProfiles in a Gen3 world, we need to know the
identity of the data from the data and not the filename. We should provide
the identity at construction time, so BuildFiberProfilesTask.run needs to
know the identity. To accomodate this cleanly, made all arguments besides
the exposure keyword arguments, which changed matching changes in several
places.