PfsConfig: get pfsDesignId,visit0 from header
It was never a good idea to rely on the filename to provide information
we should have gotten from the contents of the file, and Gen3 disabused us
of this poor practise since it uses different filename templates.

The pfsDesignId and visit0 are now in the header, so there's no need to
parse the filename except to provide backward compatibility.