add PfsTargetSpectra
It's inefficient for the pipeline to write one file per target, and it
complicates the Gen3 impementation (the registry would need to contain
the target information from the pfsConfig). Instead, we will have the
fluxCalibrate and coaddSpectra stages write files containing multiple
spectra and we will export those into pfsSingle and pfsObject files
later on.

PfsTargetSpectra is a collection of single spectra (like PfsObject),
indexed by target. It is defined here instead of in the datamodel package
because no-one outside the pipeline should ever have to deal with it.