add Gen3 utilities
In the Gen3 PipelineTask.runQuantum method, users are provided with
data references for each of the connections, but connections with similar
or even the same dimensions there is no guarantee that the data references
are aligned (i.e., they cannot be assumed to be ordered in the same way),
so simple zip iteration doesn't do what one would naively expect. This
contribution includes functions zipDatasetRefs and readDatasetRefs that
provides the alignment.

This also includes functions to ingest a pfsConfig into the datastore.
We could use "butler ingest-files", but this also reads the pfsConfig to
provide a list of tracts and patches for the registry. This is important
for the link between individual exposures and groups of coadded spectra.