FiberTraceSet.extractSpectra: accumulate mask for significant pixels
Choosing how to treat mask bits during spectrum extraction is a lot
like choosing how to treat mask bits during image coaddition. Both
involve combining multiple values with masks into a single value with
mask. It's obvious that we don't want to include bad values in the
calculation of the output value. The question is whether we want to
include bad masks in the calculation of the output mask. For
coaddition, we decided a long time ago that we don't want to include
them, because they can end up everywhere in the coadd, and they don't
relate to anything in the coadded image. For this reason, I've been
deliberately ignoring the bad mask planes when we extract
spectra. However, spectral extraction extends only over a relatively
small number of pixels (as opposed to several tens or even hundreds of
pixels contributing to each output pixel in a deep coadd), and those
pixels are often intimately related, so we're going to do something
different.

Changed extraction to accumulate (i.e., OR together) mask when
a pixel's fractional contribution (i.e., fiberTrace value divided
by the sum of the fiberTrace for that row) is in excess of a new
parameter, minFracMask. This parameter currently defaults to zero.

This will allow us to identify spectral pixels that have been
affected by cosmic rays and other effects.