convert pfsMerged fluxes to counts/nm
We previously wanted pfsMerged in units of counts (to match pfsArm), but
decided that counts/nm was more useful.

Changed MergeArmsTask to produce pfsMerged fluxes in units of counts/nm,
and then because the flux calibration is based on that.

No change to the fluxCalibrate function is required, because
we calibrate the normalised flux, which has not changed (because
changes to the flux are offset by changes to the normalisation).