Spectrum, SpectrumSet: add python typing
Adding stub files allows VSCode to do more type support for these classes
which are based in C++ and extended in python.
Adding some types in the python class extension is useful for getting
types within the *Contined.py file, but because mypy doesn't like us
extending the class, the stub files are necessary to get types when
dealing with other files.