fitContinuum: refactor to use common wavelength bins
The biggest problem with the continuum measurement has been that different
fibers have the knots at different wavelengths, because the knots have been
set independently for each fiber, in pixels. We need a consistent set of
knot wavelengths for all fibers to use. We do this by generating a set of
bins in wavelength space, and using the same bins for each fiber.

In the process, added support for operating on a PfsFiberArraySet and
PfsFiberArray instead of just Spectrum and SpectrumSet, and added python
typing information liberally.