refactor FitContinuumTask and add FitModelContinuumTask
We want to do continuum subtraction using a model constructed
from PFS observations (viz., SUNSS observations of the moon)./
It would be best if that has the same API as the regular
continuum subtraction, so we can choose to use either.
Refactored FitContinuum class into a base class
(BaseFitContinuumTask) and implementation (FitSplineContinuumTask)
and added FitModelContinuumTask and associated model
handling. The run method now (optionally) receives a
VisitInfo so we can set the zenith distance for the model-based
subtraction, and the _fitContinuumImpl operates on a PfsFiberArray
so it has access to the wavelength.