symmetricTridiagonal: deal with singular matrix
The matrix for spectral extraction can be singular if two fibers rely on
the same single pixel.

This manifests in solveSymmetricTridiagonal as denominator=0, which leads
to all values in the solution being NaN, due to the back-propagation.
Instead, when we get denominator=0, treat that element like it doesn't have
any data (set diagonal=1, offDiag=0, answer=0) and set the solution for
that element to NAN.