monkey-patch astropy's lazyproperty to avoid deadlocks with multiprocessing
astropy.utils.decorators.lazyproperty is used in a variety of places in
astropy, most notably for PFS in reading FITS tables. In astropy 4.3 (and
continuing to at least 5.1), the implementation of lazyproperty contains
thread locks which can break multiprocessing (under certain conditions).
The problem is that the fork used to generate new processes copies the lock
wholesale (see, e.g., https://stackoverflow.com/a/45889668/834250), and it
becomes possible to deadlock.

This commit adds a monkey-patch to astropy to use a pre-4.3 version of
lazyproperty, to avoid the risk of deadlocking.