ConstantFocalPlaneFunction: respect masks
We were simply taking the median, without respecting the mask, which can
allow NANs to proliferate. Unfortunately, it seems that numpy's MaskedArray
doesn't respect masks when doing a median either, so we use our own
implementation. This makes the code look a bit more complicated, but it
works.

We also need to respect the masks when summing. Setting the weights to zero
is insufficient, since it doesn't stop NANs.