table: convert to using PfsTable for I/O
Moved I/O code to datamodel, so now we need to piggyback on that. Changed
so that the Table class is built on top of the pfs.datamodel.PfsTable
class, which provides the schema and I/O functions that we delegate to.
Table uses the PfsTable schema to set up the properties and a dynamically
created RowClass. We use this scheme instead of inheritance because we
want to base Table on pandas.DataFrame instead of numpy.ndarray (Table
needs to be performant, whereas PfsTable just needs to do I/O).

In the process, changed ArcLineSet so that it uses "flux" and "fluxErr"
instead of "intensity" and "intensityErr". ReferenceLineSet continues to
use "intensity", since those numbers could be anything, whereas
ArcLineSet's intensity measurements are actually flux that we measure.