focalPlaneFunction: refactor to allow I/O by classes in datamodel
Pushing I/O down to datamodel requires a bit of a redesign. We want the I/O
only in the datamodel classes; functional code belongs in drp_stella. But
I want to avoid multiple inheritance because it's annoying. So instead of
inheriting from the datamodel class, compose it and forward most attribute
lookups there. So the datamodel class holds the data, the drp_stella class
holds the datamodel class and uses that to do the functional parts.

The refactored code has been reformatted with black and is mypy-clean.