use new PfsRepairTask to subtract traces when finding CRs
Traces (especially when bright) cause multiple problems when
finding CRs: they can be sharp and hence identified as CR
false-positives, and they increase the background making
true CRs harder to find. Subtracting the traces using a
median of a short segment of the column around each pixel
greatly improves the quality of the CR masking.