add script to export Gen3 products
We copy (or link) files into a directory structure that we can
send to users. In particular, we need to split the conglomerated
spectra into individual files for each spectrum. That's not
something we can do easily with the Gen3 registry, as it would
require having to ingest every objId into the registry, making
it bigger than we would desire.

The script uses asyncio in an attempt to hide the filesystem
overheads. This could be stripped out fairly easily, and/or
we could add parallelism to make things go faster (if the
filesystem is up to it).