centroidLines: remove traces before measuring centroids
We can't remove continuum using the fiber profiles, because in the general
case we don't know exactly where the traces are (which is why we want the
line centroids). But leaving the traces in results in inaccurate centroids
and centroid errors, so we need to get rid of the traces. Using the same
method as for cosmic ray finding appears to do the job.