fitDetectorMap: reject blended lines
We use the new getExclusionZone function to exclude blended lines.
This is separate from any exclusion zone that gets applied in the
ReadLineListTask: lines aren't flagged, which means they're not
automatically tossed downstream.

The exclusion radius is specified in pixels, which means it has a different
size in nm for different arms; this will be especially important with the
medium-resolution grating. This requires providing the dispersion to the
getGoodLines method.