fitDetectorMap: add option to reject bad reference lines
Often a particular reference line will be bad for all fibers, because
there's something going on that's not reflected in the linelist. Added an
option to detect and reject these for all fibers. This is disabled by
default as I've found it doesn't make much difference, but it may be useful
in the future.