Change filter names
TargetDB uses different filter names from what drp_stella has used before.
This change adjusts the latter to the former.

Because HSC filter names have not yet been used by targetDB,
discussion was held as to what to call HSC-R and HSC-R2.
We decided to call HSC-R "r_old_hsc" and HSC-R2 "r2_hsc" respectively
for the following reasons: Because they call HSC-R2 "r" informally,
"r_hsc" would be confusing (It is humans that prepare target records.)
"r1_hsc" would not be a good name for "HSC-R"
because there is "ENG-R1" filter, which is different from HSC-R and HSC-R2.

HSC-I and HSC-I2 are named "i_old_hsc" and "i2_hsc" by the same rule.

The integration test and the weekly test will be broken
after this change. To fix this problem, run
`PIPE2D-1110-change-filter-names.py`.
There is a link to this file in PIPE2D-1110, JIRA.