protect certain lines from exclusion zone
There are five bright sky lines in the blue that are important to be able
to adjust detectorMaps, but some of them can be discarded by exclusion
zones (e.g., NaI doublet, and one of the OI lines has nearby faint OH
lines). We add ReferenceLineStatus.PROTECTED for these cases, and have
getExclusionZone preserve those. We also allow NOT_VISIBLE lines to not
have an exclusion zone. These changes require passing the status array
to getExclusionZone.