reduceExposure: copy results of photometry
reduceExposure produces only one list of lines (the Gen3 code
has separate centroids and photometry line measurements), so
we need to copy the results from the photometry over the temporary
photometry done as part of the centroiding.
What we had been doing is dumping the traces, photometering
the lines and then keeping that; but that doesn't give us the
traces in the result.