PfsConfig: replace visit0 with visit
The 'visit0' concept (where the visit for a pfsConfig file is
the first visit with that configuration, rather than the actual
visit used) is being retired in favour of just providing the
visit. The DRP never really used the visit0 concept, so the
changes are just API changes of visit0 to visit.