centroidTraces: lower default S/N threshold to 20
With the threshold set at 50, we only detect really bright
traces. With the value at 20, we detect modestly bright
traces too, giving us lots more statistical handle for the
detectorMap adjustment.

This means that there's more noise in the trace positions,
so need to adjust the test, which was simply checking that
all positions are less than a threshold. Now testing the
RMS, and tightened down the limits appropriately.