add MultipleDistortionsDetectorMap
This allows multiple layers of distortions, which is more flexible than
simply refitting the lowest-order coefficients of a single distortion. It
will also allow adding additional distortion types in the future, such as
an empirical distortion to model the tree rings.

As part of this, expanded DoubleDistortion to support I/O within the FITS
file holding the MultipleDistortionsDetectorMap.

MultipleDistortionsDetectorMap is now the default detectorMap class coming
out of fitDetectorMap and adjustDetectorMap, but previous detectorMap
classes should continue to work fine.

Removed the detectorMap methods removeLowOrder and merge, since they
are no longer needed.