rename BaseDistortion to Distortion
For the same reason we renamed BaseDetectorMap previously: we're going to
make Distortion into something we can work with independently, and the user
doesn't need to know it's a base class.

Split BaseDistortion into Distortion, which defines the interface, and
AnalyticDistortion, which does the CRTP stuff as the foundation for the
implementations in DetectorDistortion (which should disappear eventually)
and DoubleDistortion. I can imagine later having an empirical distortion to
model the tree rings.