DoubleDistortion: fit common scale factor when fitStatic=False
We want to fit for a common scale factor (in both x and y, for both CCDs;
due to small focus changes), which allows us to model the wavelength
distortion when we only have continuum. This requires a non-linear fit,
so I'm using Minuit2, copied from the fitLine implementation. This might be
a bit slower, but it only runs at low order, and so shouldn't be too awful.