DoubleDistortion: fit scale+rotation for left+right
I was fitting for a common scale factor for x/y and left/right,
but that appears to be insufficient. Changed to fitting a scale
and rotation that is different for the left and right CCDs.
This allows the CCDs to move slightly differently.
In one example, this reduced the xSoften from 0.08 to 0.03,
which is where it was before I started fitting for the scale,
but now with the advantage that the wavelength solution is
constrained.