estimateRadialVelocity: Change `findMethod` to "quadratic"
Previously, we found the peak position of the cross correlation function
by fitting a Gaussian to the c.c.f. Because exceptions were sometimes
raised by the optimizer, we change the peak-finding method to quadratic
interpolation. Because radial velocities found by the latter method
have large variance when `searchStep` is small, we change it from 5 km/s
to 10 km/s (This step should not be wider than the smallest possible
LSF). In addition, we lower S/N used in the unit test according to
true observed spectra.