add RotScaleDistortion
Distortion consists of translation, rotation and scale for
left and right detectors separately. Unfortunately, the
scale factor for x and y are different (the data seems to
demand it), which means we really need wavelength references
in the frame in order to get good wavelength solutions.

Cleaned up the FitDistortedDetectorMapTask API for fitting
the distortion, so it's easy to override.