quartiles: add support for calculating median/quartiles without mask
The calculateQuartiles and calculateMedian functions only accepted both
values and a mask array. Modified both to work with either values+mask or
values only. This is done by using a Partitioner object which hides from
the algorithmic code the details of whether there is a mask involved.

Added a test for both these functions.