adjust to inclusion of reduction notes in spectral products
This means we can no longer use PfsFiberArray or PfsFiberArraySet directly,
since they don't have the NotesClass class variable defined; so we use
PfsSingle or PfsObject as appropriate.

In order to extract a PfsFiberArray from the PfsTargetSpectra, we need to
know the schema for the notes, which requires having the PfsTargetSpectra
know which type of PfsFiberArray it contains, and a table version of the
notes. For this reason, we now subclass PfsTargetSpectra as
PfsCalibrateddSpectra and PfsObjectSpectra, which have the class variables
set appropriately.