add black spot correction
The black spots have a zone of influence (penumbra) larger than the area
where they completely shadow the fiber (umbra). The throughput loss appears
to be linear in the distance from the black spot. I have measured a
correction from fitting a bunch of lines in a couple of different
exposures, and have added this correction to the 'reduceExposure' stage.
The correction is applied to the normalisation, so that we don't modify
the flux values that we extract from the image.

The correction is written to the new 'notes' feature of the spectra.