FiberTrace: normalise trace in constructImage
The extracted spectra have the flux in counts. We should be
able to feed this into SpectrumSet::makeImage, which uses
FiberTrace::constructImage, and get something like the original
image. However, FiberTrace::constructImage is applying the
trace normalisation, so it requires that the input spectra be
normalised. That's an extra step (and that extra step is now
subject to error, since the extracted spectrum's normalisation
gets changed by the black spot correction, and so it would no
longer match the trace normalisation), which should be
eliminated.