FiberTraceSet: use sparse matrix for extraction
When extracting spectra, the contributions for neighbouring fibers overlap
and so we've been using a symmetric tridiagonal matrix solver. However,
that only deals with overlaps between immediate neighbours. With wider
fiber profiles, we can get overlaps between fibers that aren't immediate
neighbours. We could probably find a penta-diagonal matrix solving
algorithm, but we already have a sparse matrix solver and then we don't
need to scale it up again later to a septa-diagonal solver.

We still use the tridiagonal solver to get the covariances, since the
inversion of a sparse matrix is (in general) not sparse, so this provides
a fast approximation.