fitFluxCal: limit fibers to those we have available
Fixes 'KeyError: NO_DATA' from running makeFluxTable with an
empty list of spectra. We were iterating over fibers which
didn't have any data, because user was specifying
spectrograph=1^3 in the --id.

Not going to fix the root cause, as this is Gen2-specific, and
we're hoping to dump that soon.