Add script to construct fiber profiles from simultaneous fit
Fibers can't always be hidden with 100% effectiveness, and some fibers
can't ever be hidden. In order to get a good set of fiber profiles in the
face of these difficulties, we need to fit the fiber profiles
simultaneously, so we're not getting flux from one fiber profile affecting
another. This commit adds a new script, reduceProfiles.py, with associated
support, to accomplish this.