fitSwathProfiles: robustify profile fitting
Switching to a conjugate gradient solver works around the matrix being
close to singular. Then we can't use a self-adjoint view, so we need to
provide the whole matrix.
That's now good enough and fast enough that we can use the linear
interpolation code path, so enabled that.

Cleaned up the interation ranges.