reduceExposure: add background subtraction
As evidenced by signficantly-positive counts on the side edges
and the strip down the middle, there's some scattered light that
needs to be subtracted off. This is not easy, as the fibers
don't provide many places where this can be measured. We end up
with just the sides and the strip in the middle, and so fit
only a 2nd order polynomial in x, but a much higher order in
y.