fluxModelSet: Read spectra from a zip file
Reading very many spectra speeds up when they are stored in an uncompressed
zip file. With fluxmodeldata-ambre-20220714-full, speedup of the integration
test by 50 min/visit was observed. It is essential to use a zip file
instead of, say, a multi-extension FITS file because astropy would use up
all the memory during an iteration over the HDUList.