fitPfsFluxReference: Downsample models before continuum fitting
Flux models' sampling interval is very fine so as to trace their narrow
LSF, but they can be downsampled after a wider LSF is convolved with them.
Downsampling the models speeds up continuum fitting, which is the very hot
spot of FitPfsFluxReferenceTask. We still dare not downsample the models
with the same sampling interval as observed spectra's because the models
have to be resampled again, after continuum fitting, due to Doppler shift.
We therefore introduce a new config parameter modelOversamplingFactor
defaulting to 2.0, which implies that the models are downsampled with
sampling interval twice as fine as observed spectra's.