SparseSquareMatrix: use map-based triplets
The vector-based triplets requires us to store all the additions to be
made later. This results in a huge memory usage (probably because we're
drastically under-estimating how much memory we need to start with, and
then increasing it piecemeal). Instead, make the additions as we go,
and give Eigen just a single value for each non-zero matrix element.
This drastically decreases the memory usage (70% --> 1.5% of pfs-usr02
at Hilo).