fluxModelSet: Removed unintentional permanent reference
When we prepend @functools.cache to a member function and call it,
`self` in the arguments gets stored in the wrapper owned by the class type.
This reference to the class instance will be kept until the class type is
deleted, which will never happen. Hence resource leak is reported by
`MemoryTestCase`.