constructFiberFlat: smooth spectra used for normalisation
We normalise the input images by the spectrum. This gives us
the deviation from what is expected for pixels in the spatial
dimension. However, this doesn't measure any deviation from
expected for pixels in the spectral dimension. Smoothing the
spectra (we use a median filter) allows us to sample the
deviations in the spectral dimension.

This is very important for the n-band, where there are
strong deviations from flatness. Without this change, the
flat-field looks like the Laplacian of the flat-field, and
not the flat-field itself.