fitPfsFluxReference: Fit models using Powell's method
Previously, fitPfsFluxReference fitted models to spectra by evaluating
chi^2 for every precomputed model template. After this change,
fitPfsFluxReference uses Powell's method (by default) to fit models.

With this change, fitPfsFluxReference.py runs 3x faster (if the number
of FLUXSTD in a visit is 100). The quality of the output seems slightly
better nonetheless for the possibility of falling into a local minimum,
to my surprise. It may be that the sub-grid peak finder in the previous
method was bad, or that LSF approximation for speedup was problematic,
though these are mere guesses.

The previous method can be restored with
`config.minimizationMethod='brute-force'`.