remove Jacobian corrections
Arnaud pointed out that the use of the Jacobian in interpolation
is incorrect, since we have already corrected for the dispersion.

After correcting for the sampling differences (by dividing by
the dispersion in nm/pixel), the fluxes are in units of
electrons/nm, which is what we want. Applying a Jacobian means
we care about the flux per pixel, as if the flux value is
integrated over the spectral pixel; but that's not what we want.
The flux for a spectral pixel in pfsMerged is the flux (in
electrons/nm) at that spectral pixel's wavelength. The error is
analagous to what RHL has said before in the context of imaging
pixels: the value of the image at a particular position is the
value of whatever we're measuring at the center of the pixel,
not the integrated value over the pixel.

I've therefore removed all mention of Jacobians (because they
only matter when we care about the size of a pixel, and we do
not). Dividing the pfsArm flux (in electrons) by the dispersion
(nm per pixel) is sufficient to perform the desired conversion
from electrons to electrons per nm for pfsMerged, and that's all
we need.