use minimum dispersion for wavelength sampling
We have been using a dispersion of about 0.081 nm/pixel for
our interpolated products, but the blue arm has a minimum
dispersion of 0.065 nm/pixel, which means we're losing data.

Instead of hardwiring a dispersion, calculate the minimum
dispersion (in nm/pixel) and use that for the entire range.
This is not an optimal solution (which might involve fitting
a dispersion as a function of wavelength to the lower envelope
of the input dispersions), but it means we won't lose data.