buildFiberProfiles: use Gaussian approximation to measure flux when iterating
When iterating with FiberProfileSet.fromImages (used in
reduceProfiles), we've been using the measured profile to
extract the spectra that we use in the next iteration.
However, the measured profile might contain features that
don't give a good flux measurement, which means the
convergence proceeds slowly (this was especially the case
with r3 when it had an optical misalignment). Using a
Gaussian approximation to the measured profiles to extract
the spectra removes any nasty features, and greatly speeds
up convergence.