make reduceExposure more widely parallelisable
We want to run reduceExposure in parallel for multiple visits
at once. We no longer expect to run sky2d before transitioning
to the Gen3 middleware, so we can remove the limitation to
operating over all arms in a visit, and now simply parallelise
reduceExposure over all available arms.

This results in some API changes to ReduceExposureTask,
including the naming of results.