add synchronised x/y fitting and RotScaleDistortion
Because the traces are curved, the position of the continuum in the
spatial dimension allows us to constrain the distortion in the spectral
dimension as well. This requires simultaneous fitting for the x and y
distortions, and converting the y distortion (for Trace measurements only)
to an x distortion using the slope of the trace. Added this feature to
the PolynomialDistortion fitter and to the evaluation of the residuals.

To make full use of this feature, we introduce the RotScaleDistortion,
which includes only a translation, rotation and a common scale factor for
x and y (e.g., due to small focus changes), and make it the default for
adjusting the detectorMap.