fitPfsFluxReference: Discard FLUXSTD when its param is out of range
If the guessed parameter of a spectrum is out of the parameter domain
of the models we have prepared, we should discard the spectrum.
In a previous commit, we already took care of temperature after the
call to fitBroadbandSED(). In this commit, we take care of all
dimensions after spectrum fit.