fitDetectorMap: disable exclusion radius for reduceArc
When running reduceArc, we combine arc line measurements from many
exposures. The exclusion zone was tossing many lines that weren't
actually getting in each other's way because they were on different
exposures. Disabling the exclusion zone for reduceArc deals with
this in a simple way (an alternative would be to apply the
exclusion zone for each exposure separately, but that's a little
more complicated). I'm leaving the exclusion zone active for
adjustDetectorMap, since that's frequently used with sky lines
where we really do need to exclude lines, and in this case all
lines are on the same image so it's valid.