solveLeastSquares: allow forcing parameters to provided values
We want to be able to force parameters of a fit to user-provided values.
To this end, we can now supply an array of booleans indicating whether the
parameter should be fixed, along with an array of values for the parameters
(which default to zero if not provided). If a parameter is fixed, the
matrix solver removes it from the equation, and provides it directly in the
solution.