setCalibHeader: clobber any existing CALIB_INPUT_nnn headers
We record the list of inputs in the CALIB_INPUT_nnn headers,
but didn't consider whether the header might already have
such headers. If it does, and the list is longer than the
new list, then we would leave some of the old headers,
confusing anyone who reads the list later.
Instead, we remove all the CALIB_INPUT_nnn headers before
putting the new ones in.