fitPfsFluxReference: Use a 1D prior instead of 4D one
We make a 4D (Teff, log(g), metal, alpha) PDF from broad-band
fitting, and use it as the prior in spectrum fitting.
It has turned out that its interpolator is one of bottlenecks.
Because only Teff can be determined by broad-band fitting,
we can marginalize other dimensions to get a 1D prior.
1D interpolators are much faster than 4D ones.