FiberTrace: add compile-time option for linear interpolation
My experiments indicate that linear interpolation is more accurate than
spline interpolation when constructing the FiberTrace from profiles.
However, this is counter-intuitive and has not yet been proved to the
satisfaction of the group, so this option is currently disabled, but I've
left it as a compile-time option so we can play with it later.