buildFiberProfiles: be more careful with fluxes and detecting bad fibers
It's very important to get the flux correct: we can get bad profiles or
even bad matrix equations if we blindly measure the flux. We need to
identify bad flux measurements (using the mask; setting minFracMask helps)
and interpolating over them so our flux measurements are reasonable (we
can't have masked fluxes going into the matrix equation, because they
would propagate across fibers through the cross terms, so better to just
guess what the flux should be, and then if we're wrong the pixels will get
masked rather than the complete fiber).

Also, sometimes the pfsConfig is inaccurate and a fiber is not actually
illuminated when we think it is. We can detect those (as having low flux)
and give them a dummy profile that allows the process to complete cleanly.

Also fixed a bug where the fluxes were being provided in the wrong order:
the extracted spectra may be in a different order than we expect.