reduceProfiles: use dot-roach dark subtraction
The "black spots" aren't perfectly black, allowing a bit of light through.
Measuring the flux of these (which we need in order to measure the profile)
when we don't yet know the profile is difficult (and accurate fluxes are
essential for measuring the profile). Instead, we subtract a "dot-roach
dark", which is an exposure with all fibers hidden, which removes all the
flux from the hidden fibers. Then we can measure the profiles for the
exposed fibers, and merge the profiles measured this way from multiple
groups of fiber configurations.