NormalizedPolynomial: fix __init__
An array with a single element as the first argument in the constructor
was getting the wrong overload: the array was being interpreted as a
scalar and being cast to an integer. Fixed by declaring the overloads in
the correct order.

Also added an import of lsst.afw.math, since it seems to be required to
load the base class in at least some cases.