add construction and use of fiberNorms
The relative flux normalisation of fibers can change, especially when the
instrument is installed and connections are made. The new fiberNorms
calib product is intended to correct for this.

The PfsFiberNorms carries coefficients for polynomials, but we're only
using a constant value at the moment, at least until we can verify that we
need more and they can be calculated robustly.