mergeArms,coaddSpectra: mask only NO_DATA extracted pixels in coaddition
CR, SAT, BAD_FLAT etc masks don't mean that the extracted pixel
is necessarily bad, only that a significant pixel in the image
had that mask (but it was ignored in the extraction). We therefore
don't want to clip those in the coadd, as we would be tossing data
unnecessarily.

BAD_SKY and BAD_FLUXCAL are different: they don't come from the
extraction, so it's legitimate to toss those.