fitFluxCal.py: Address problems arising from absent arms.
If an arm is absent, spectrum in its wavelength range is not available.
Because we compare broadband fluxes with integrations of spectra,
sampling points absent from the integration ranges lead to erroneous
chi^2.

To address this problem, we throw away broadband fluxes if the bands
are hardly covered by the observed spectra. In addition, we interpolate
or extrapolate the observed spectra so that they will cover entire
bands of broadband filters.