fitFluxCal: Use wavelength-independent polynomial
We have fit a polynomial dependent on (x, y, \lambda) to the flux
calibration vector (observed flux / model flux). But the dependence
on \lambda makes calibration vector quite wrong at the red end of
spectra. We therefore use a wavelength-independent polynomial
instead. (In fact, the polynomial is a modification factor by which
to multiply the raw calibration vector. The flux calibration vector
remains wavelength-dependent.)