add LayeredDetectorMap
This detectorMap consists of layers of corrections on top of the optical
model: slit offsets, optical distortions, and divided detectors. It allows
for fibers to fall between the chip gaps, rather than treating the divided
b,r,m detectors as contiguous.

Adopting this as the standard detectorMap class going forward.