refactor fiberNorms measurement
constructFiberNorms.py was used to measure fiberNorms from raw data. This
is more cumbersome than measuring from extracted spectra. Furthermore, we
want the fiberNorms to be measured for an arm using all spectrographs.
measureFiberNorms.py replaces constructFiberNorms.py. It uses pfsArms
(from running reduceExposure.py) to build fiberNorms and a plot of
fiberNorm values on the focal plane. The fiberNorms can then be ingested
into the calibs, if desired. This allows faster checking of the status of
the instrument.